Refactor the InputCurrency to use dynamic Id and use util to parse the input value
