+++ src/util/EventParserUtil.js
@@ -1,0 +1,11 @@
+class EventParserUtil {
+  static parseEventValueToNumberIfValueExist = event => {
+    if (event && event.target && event.target.value) {
+      return parseFloat(event.target.value);
+    }
+  
+    return "";
+  }
+}
+
+export default EventParserUtil;
